refactor(currency): Currency conversion in system was not solid. So the currency types and their usage is become standart.

+++ Backend/src/SoftPMS.Application/Features/Payrolls/Queries/GetEmployeePayrollSlips/GetEmployeePayrollSlipsQueryHandler.cs
@@ -26,7 +26,7 @@ public sealed class GetEmployeePayrollSlipsQueryHandler(
                 p.NetSalary,
                 p.SalaryTypes,
                 p.IssueDate,
-                p.LineItems.Select(li => new PayrollSlipLineItemDto(li.Id, (int)li.ItemType, li.Description, li.Amount, (int)li.Currency)).ToList()))
+                p.LineItems.Select(li => new PayrollSlipLineItemDto(li.Id, li.ItemType, li.Description, li.Amount, li.Currency)).ToList()))
             .ToListAsync(cancellationToken);
     }
 }
